Added unicode flag to the ES6 `RegExp.prototype[Symbol.match]` test
This makes it pass in any browser with or without unicodeSets support. Browsers that check the individual properties (in compliance with ES6) instead of flags (in compliance with later versions of the spec) check unicode first, and check unicodeSets if unicode is not truthy. Current versions of V8 in particular are affected, after adding support for unicodeSets, but still not looking up flags.